# docs/api-contracts/admin-module-openapi.yaml
...
openapi: 3.0.3
info:
  title: Admin Module API
  description: |
    Admin Module — управляющий модуль системы Artstore.

    **Admin Module не является auth-сервером.** Аутентификация и выдача JWT
    полностью делегированы внешнему Identity Provider (Keycloak). Admin Module
    получает JWT от API Gateway, проверяет claims (роли, scopes) и принимает
    решения по авторизации для своих endpoints.

    ## Аутентификация

    Все запросы (кроме Health) проходят через API Gateway, который:
    - Проверяет подпись JWT через JWKS Keycloak
    - Проверяет expiration
    - Форвардит claims в заголовках к Admin Module

    ### Admin Users (H2M)
    Аутентификация через Keycloak OIDC (Authorization Code + PKCE).
    JWT содержит `sub`, `realm_access.roles`, `groups`.

    Роли определяются маппингом Keycloak groups → Artstore roles:
    - `artstore-admins` → `admin` (полный доступ)
    - `artstore-viewers` → `readonly` (только чтение)

    Admin Module может **дополнить** (но не понизить) роль через
    локальные role overrides.

    ### Service Accounts (M2M)
    OAuth 2.0 Client Credentials flow через Keycloak.
    JWT содержит `client_id`, `scope`.

    Scopes:
    - `files:read` — чтение метаданных и скачивание файлов
    - `files:write` — загрузка, обновление и удаление файлов
    - `storage:read` — чтение информации о Storage Elements
    - `storage:write` — управление SE (sync, mode transition)
    - `admin:read` — чтение административных данных
...

    ## Shared PostgreSQL
...

...

...

...
servers:
...
paths:
...
  /api/v1/files:
    post:
      tags: [files]
      summary: Регистрация файла
      description: |
        Регистрирует файл в реестре после загрузки на SE.
        Вызывается Ingester Module.

        Доступно SA с scope `files:write`.
      operationId: registerFile
      security:
        - bearerAuth: [files:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileRegisterRequest"
      responses:
        "201":
          description: Файл зарегистрирован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileRecord"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Файл с таким ID уже зарегистрирован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      tags: [files]
      summary: Список файлов
      description: |
        Возвращает пагинированный список файлов из реестра.

        Доступно: SA с scope `files:read`, роли `admin`, `readonly`.
      operationId: listFiles
      security:
        - bearerAuth: [files:read]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: status
          in: query
          description: Фильтр по статусу
          schema:
            type: string
            enum: [active, expired, deleted]
        - name: retention_policy
          in: query
          description: Фильтр по политике хранения
          schema:
            type: string
            enum: [temporary, permanent]
        - name: storage_element_id
          in: query
          description: Фильтр по Storage Element
          schema:
            type: string
            format: uuid
        - name: uploaded_by
          in: query
          description: Фильтр по загрузившему
          schema:
            type: string
      responses:
        "200":
          description: Список файлов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileRecordListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/files/{file_id}:
    get:
      tags: [files]
      summary: Метаданные файла
      description: |
        Возвращает метаданные файла из реестра.

        Доступно: SA с scope `files:read`, роли `admin`, `readonly`.
      operationId: getFile
      security:
        - bearerAuth: [files:read]
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "200":
          description: Метаданные файла
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      tags: [files]
      summary: Обновление метаданных файла
      description: |
        Обновление метаданных файла в реестре (description, tags, status).

        Доступно: SA с scope `files:write`, роль `admin`.
      operationId: updateFile
      security:
        - bearerAuth: [files:write]
      parameters:
        - $ref: "#/components/parameters/FileId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileRecordUpdate"
      responses:
        "200":
          description: Метаданные обновлены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileRecord"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [files]
      summary: Soft delete файла
      description: |
        Помечает файл как удалённый (status → `deleted`).
        Физическое удаление на SE выполняется GC на SE.

        Доступно: SA с scope `files:write`, роль `admin`.
      operationId: deleteFile
      security:
        - bearerAuth: [files:write]
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "204":
          description: Файл помечен как удалённый
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # =========================================================================
  # IdP Status (2 endpoints)
  # =========================================================================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT от Keycloak. Валидация подписи — на API Gateway.
        Admin Module проверяет claims (роли, scopes).

        Для Admin Users: `realm_access.roles`, `groups`
        Для Service Accounts: `scope`

  # =========================================================================
  # Параметры
  # =========================================================================
  parameters:
    Limit:
      name: limit
      in: query
      description: Количество записей на странице
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
    Offset:
      name: offset
      in: query
      description: Смещение от начала списка
      schema:
        type: integer
        minimum: 0
        default: 0
...
    FileId:
      name: file_id
      in: path
      required: true
      description: UUID файла
      schema:
        type: string
        format: uuid

  # =========================================================================
  # Переиспользуемые ответы
  # =========================================================================
  responses:
    Unauthorized:
      description: Отсутствует или невалидный JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: "Требуется аутентификация"
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: FORBIDDEN
              message: "Недостаточно прав для выполнения операции"
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: NOT_FOUND
              message: "Ресурс не найден"
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INTERNAL_ERROR
              message: "Внутренняя ошибка сервера"

  # =========================================================================
  # Схемы данных
  # =========================================================================
  schemas:
...
    FileRecord:
      type: object
      description: Запись файла в реестре
      required:
        - file_id
        - original_filename
        - content_type
        - size
        - checksum
        - storage_element_id
        - uploaded_by
        - uploaded_at
        - status
        - retention_policy
      properties:
        file_id:
          type: string
          format: uuid
        original_filename:
          type: string
          example: photo_landscape.jpg
        content_type:
          type: string
          example: image/jpeg
        size:
          type: integer
          format: int64
          example: 5242880
        checksum:
          type: string
          description: SHA-256
          example: a1b2c3d4e5f67890abcdef1234567890abcdef1234567890abcdef1234567890
        storage_element_id:
          type: string
          format: uuid
        uploaded_by:
          type: string
          example: sa_ingester_a1b2c3
        uploaded_at:
          type: string
          format: date-time
        description:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
          example: ["logo", "draft"]
        status:
          type: string
          enum: [active, expired, deleted]
          example: active
        retention_policy:
          type: string
          enum: [temporary, permanent]
          example: permanent
        ttl_days:
          type: integer
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FileRegisterRequest:
      type: object
      description: Регистрация файла (от Ingester)
      required:
        - file_id
        - original_filename
        - content_type
        - size
        - checksum
        - storage_element_id
        - retention_policy
      properties:
        file_id:
          type: string
          format: uuid
        original_filename:
          type: string
        content_type:
          type: string
        size:
          type: integer
          format: int64
        checksum:
          type: string
        storage_element_id:
          type: string
          format: uuid
        description:
          type: string
          maxLength: 1000
        tags:
          type: array
          items:
            type: string
        retention_policy:
          type: string
          enum: [temporary, permanent]
        ttl_days:
          type: integer
          minimum: 1
          maximum: 365
          description: Обязателен для retention_policy=temporary
    FileRecordUpdate:
      type: object
      properties:
        description:
          type: string
          maxLength: 1000
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, expired, deleted]
      minProperties: 1
    FileRecordListResponse:
      type: object
      required:
        - items
        - total
        - limit
        - offset
        - has_more
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FileRecord"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean

    # -----------------------------------------------------------------------
    # IdP
    # -----------------------------------------------------------------------
...
    ErrorResponse:
      type: object
      description: Стандартный формат ошибки (единый для всей системы Artstore)
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: |
                Машиночитаемый код ошибки:
                - `VALIDATION_ERROR` — некорректные входные данные
                - `NOT_FOUND` — ресурс не найден
                - `UNAUTHORIZED` — требуется аутентификация
                - `FORBIDDEN` — недостаточно прав
                - `CONFLICT` — конфликт (дублирующийся ресурс)
                - `SE_UNAVAILABLE` — SE недоступен
                - `IDP_UNAVAILABLE` — Keycloak недоступен
                - `INTERNAL_ERROR` — внутренняя ошибка
              example: NOT_FOUND
            message:
              type: string
              description: Человекочитаемое описание
              example: Ресурс не найден